apps/category/urls.py
from django.db import router
from .views import CategoryViewSet
from rest_framework.routers import DefaultRouter



router = DefaultRouter()
router.register('category', CategoryViewSet)


urlpatterns = []
urlpatterns += router.urls
